Did some clean up. Need to stop OdometryGlobalCoordinatePosition on OpMode exit.
