+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/robotAttachments/Subsystems/LEDs.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.src.robotAttachments.Subsystems;
+
+import org.firstinspires.ftc.teamcode.src.Utills.ThreadedSubsystemTemplate;
+
+public class LEDs extends ThreadedSubsystemTemplate {
+    protected void threadMain() {
+
+    }
+}
